ai-services/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Pillow==11.1.0
pandas==2.2.3
numpy==2.3.1
opencv-python==4.10.0.84
google-generativeai==0.8.5
pillow-heif==1.1.0
selenium==4.28.1
beautifulsoup4==4.13.3
schedule==1.2.2
python-dotenv==1.0.1
numpy==1.24.3